api/pyproject.toml
[project]
name = "cyclejournal-api"
version = "0.1.0"
description = "CycleJournal Backend API - GCP Cloud Run + Firestore"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.5.0",
    "google-cloud-firestore>=2.19.0",
    "anthropic[vertex]>=0.42.0",
    "pyjwt[crypto]>=2.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
